create statusCommand
